+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.18.0
 
+- implement [@Use()](https://egomobile.github.io/node-http-server/modules.html#Use)
 - replace [mime-types](https://www.npmjs.com/package/mime-types) with [mrmime](https://www.npmjs.com/package/mrmime)
 - `npm update`s
 - (bug-)fixes